run_analyzer/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
openai>=1.30.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
